+++ examples/extras/curves/cubic_bezier_curve.cpp
@@ -7,7 +7,7 @@ using namespace threepp;
 
 int main() {
 
-    Canvas canvas(Canvas::Parameters().antialiasing(8));
+    Canvas canvas("CubicBezierCurve", {{"antialiasing", 4}});
     GLRenderer renderer(canvas);
     renderer.setClearColor(Color::aliceblue);
