+++ examples/loaders/obj_loader.cpp
@@ -19,7 +19,7 @@ void createAndAddLights(Scene& scene) {
 
 int main() {
 
-    Canvas canvas{Canvas::Parameters().antialiasing(8)};
+    Canvas canvas("OBJ loader", {{"antialiasing", 4}});
     GLRenderer renderer(canvas);
 
     auto scene = Scene::create();
