+++ examples/loaders/stl_loader.cpp
@@ -5,7 +5,7 @@ using namespace threepp;
 
 int main() {
 
-    Canvas canvas{Canvas::Parameters().antialiasing(4)};
+    Canvas canvas("STL loader", {{"antialiasing", 4}});
     GLRenderer renderer(canvas);
     renderer.setClearColor(Color::aliceblue);
