+++ examples/misc/CMakeLists.txt
@@ -1,18 +1,18 @@
 
 add_executable(key_listener key_listener.cpp)
-target_link_libraries(key_listener PUBLIC threepp)
+target_link_libraries(key_listener PUBLIC threepp glfw::glfw)
 
 add_executable(mouse_listener mouse_listener.cpp)
-target_link_libraries(mouse_listener PUBLIC threepp)
+target_link_libraries(mouse_listener PUBLIC threepp glfw::glfw)
 
 add_executable(raycast raycast.cpp)
-target_link_libraries(raycast PUBLIC threepp)
+target_link_libraries(raycast PUBLIC threepp glfw::glfw)
 
-add_executable(test_capabilites test_capabilites.cpp)
-target_include_directories(test_capabilites PUBLIC ${PROJECT_SOURCE_DIR}/src)
-target_link_libraries(test_capabilites
-        PUBLIC
-        threepp
-        PRIVATE
-        glfw::glfw
-        glad::glad)
+#add_executable(test_capabilites test_capabilites.cpp)
+#target_include_directories(test_capabilites PUBLIC ${PROJECT_SOURCE_DIR}/src)
+#target_link_libraries(test_capabilites
+#        PUBLIC
+#        threepp
+#        PRIVATE
+#        glfw::glfw
+#        glad::glad)
