+++ examples/projects/Crane3R/CMakeLists.txt
@@ -1,6 +1,6 @@
 
 add_executable(Crane3R main.cpp Crane3R.cpp)
-target_link_libraries(Crane3R PRIVATE threepp)
+target_link_libraries(Crane3R PRIVATE threepp glfw::glfw)
 target_include_directories(Crane3R PRIVATE "${PROJECT_SOURCE_DIR}/examples/libs")
 
 if (imgui_FOUND)
