+++ examples/shaders/CMakeLists.txt
@@ -1,9 +1,9 @@
 
 add_executable(raw_shader raw_shader.cpp)
-target_link_libraries(raw_shader PUBLIC threepp)
+target_link_libraries(raw_shader PUBLIC threepp glfw::glfw)
 
 add_executable(seascape_demo seascape_demo.cpp)
-target_link_libraries(seascape_demo PUBLIC threepp)
+target_link_libraries(seascape_demo PUBLIC threepp glfw::glfw)
