+++ examples/textures/texture2d.cpp
@@ -9,7 +9,7 @@ using namespace threepp;
 
 int main() {
 
-    Canvas canvas(Canvas::Parameters().antialiasing(8));
+    Canvas canvas(CanvasOptions().antialiasing(8));
     GLRenderer renderer(canvas);
     renderer.setClearColor(Color::aliceblue);
