+++ include/threepp/canvas/Canvas.hpp
@@ -2,85 +2,65 @@
 #ifndef THREEPP_CANVAS_HPP
 #define THREEPP_CANVAS_HPP
 
+#include "threepp/canvas/CanvasOptions.hpp"
 #include "threepp/canvas/WindowSize.hpp"
 #include "threepp/input/PeripheralsEventSource.hpp"
+#include "threepp/textures/Image.hpp"
 
 #include <filesystem>
 #include <functional>
 #include <memory>
 #include <optional>
+#include <queue>
 #include <string>
 #include <unordered_map>
 #include <variant>
 
 namespace threepp {
 
-    class Canvas: public PeripheralsEventSource {
+    namespace detail {
+        typedef std::pair<std::function<void()>, float> Task;
+
+        struct CustomComparator {
+            bool operator()(const Task& l, const Task& r) const { return l.second > r.second; }
+        };
+    }// namespace detail
+
+    class CanvasBase: public PeripheralsEventSource {
 
     public:
-        struct Parameters;
-        typedef std::variant<bool, int, std::string, WindowSize> ParameterValue;
-
-        explicit Canvas(const Parameters& params = Parameters());
-
-        explicit Canvas(const std::string& name);
-
-        Canvas(const std::string& name, const std::unordered_map<std::string, ParameterValue>& values);
+        explicit CanvasBase(const WindowSize& size);
 
         [[nodiscard]] const WindowSize& getSize() const;
 
         [[nodiscard]] float getAspect() const;
 
-        void setSize(WindowSize size);
-
         void onWindowResize(std::function<void(WindowSize)> f);
 
         void animate(const std::function<void()>& f);
 
-        // returns false if application should quit, true otherwise
-        bool animateOnce(const std::function<void()>& f);
-
         void invokeLater(const std::function<void()>& f, float t = 0);
 
-        [[nodiscard]] void* windowPtr() const;
+        virtual void setSize(WindowSize size) = 0;
 
-        ~Canvas() override;
+        // returns false if application should quit, true otherwise
+        virtual bool animateOnce(const std::function<void()>& f) = 0;
+
+        ~CanvasBase() override;
+
+    protected:
+        WindowSize size_;
+        std::optional<std::function<void(WindowSize)>> resizeListener_;
+
+        void handleTasks();
+
+        static std::optional<Image> loadFavicon();
+
+        virtual float getTime() = 0;
 
     private:
-        struct Impl;
-        std::unique_ptr<Impl> pimpl_;
+        std::priority_queue<detail::Task, std::vector<detail::Task>, detail::CustomComparator> tasks_;
 
-    public:
-        struct Parameters {
-
-            Parameters();
-
-            explicit Parameters(const std::unordered_map<std::string, ParameterValue>& values);
-
-            Parameters& title(std::string value);
-
-            Parameters& size(WindowSize size);
-
-            Parameters& size(int width, int height);
-
-            Parameters& antialiasing(int antialiasing);
-
-            Parameters& vsync(bool flag);
-
-            Parameters& resizable(bool flag);
-
-            Parameters& favicon(const std::filesystem::path& path);
-
-        private:
-            WindowSize size_{640, 480};
-            int antialiasing_{0};
-            std::string title_{"threepp"};
-            bool vsync_{true};
-            bool resizable_{true};
-            std::optional<std::filesystem::path> favicon_;
-
-            friend struct Canvas::Impl;
-        };
     };
 
 }// namespace threepp
