+++ include/threepp/controls/OrbitControls.hpp
@@ -10,7 +10,7 @@
 
 namespace threepp {
 
-    class Canvas;
+    class CanvasBase;
 
     class OrbitControls {
 
@@ -61,7 +61,7 @@ namespace threepp {
         // Set to false to disable use of the keys
         bool enableKeys = true;
 
-        OrbitControls(Camera& camera, Canvas& canvas);
+        OrbitControls(Camera& camera, CanvasBase& canvas);
 
         bool update();
